requirements.txt
dohq-artifactory
